Create Countelementslinkedlist.c
Counting the number of elements in a linked list is a common operation when working with linked lists. A linked list is a data structure consisting of nodes, where each node contains a value and a pointer to the next node in the sequence. To count the number of elements in a linked list, we need to traverse the list and increment a counter for each node encountered.